now do the actual key checks to unsubscribe

+++ unsubscribe.php
@@ -34,15 +34,37 @@ if (isset($_GET['user_id']) && is_numeric($_GET['user_id']) && isset($_GET['emai
 		if (isset($_GET['article_id']) && is_numeric($_GET['article_id']))
 		{
 			// check secret key
-			$db->sqlquery("DELETE FROM `articles_subscriptions` WHERE `user_id` = ? AND `article_id` = ?", array($_GET['user_id'], $_GET['article_id']));
-			header("Location: home/message=unsubscribed");
+			$db->sqlquery("SELECT `secret_key` FROM `articles_subscriptions` WHERE `user_id` = ? AND `article_id` = ? AND `secret_key` = ?", array($_GET['user_id'], $_GET['article_id'], $_GET['secret_key']));
+			$check_key = $db->fetch();
+			$check_exists = $db->num_rows();
+			
+			if ($check_exists == 1)
+			{
+				$db->sqlquery("DELETE FROM `articles_subscriptions` WHERE `user_id` = ? AND `article_id` = ?", array($_GET['user_id'], $_GET['article_id']));
+				header("Location: home/message=unsubscribed");
+			}
+			else
+			{
+				header("Location: home/message=cannotunsubscribe");
+			}
 		}
 
 		else if (isset($_GET['topic_id']) && is_numeric($_GET['topic_id']))
 		{
 			// check secret key
-			$db->sqlquery("DELETE FROM `forum_topics_subscriptions` WHERE `user_id` = ? AND `topic_id` = ?", array($_GET['user_id'], $_GET['topic_id']));
-			header("Location: home/message=unsubscribed");
+			$db->sqlquery("SELECT `secret_key` FROM `forum_topics_subscriptions` WHERE `user_id` = ? AND `topic_id` = ? AND `secret_key` = ?", array($_GET['user_id'], $_GET['topic_id'], $_GET['secret_key']));
+			$check_key = $db->fetch();
+			$check_exists = $db->num_rows();
+			
+			if ($check_exists == 1)
+			{
+				$db->sqlquery("DELETE FROM `forum_topics_subscriptions` WHERE `user_id` = ? AND `topic_id` = ?", array($_GET['user_id'], $_GET['topic_id']));
+				header("Location: home/message=unsubscribed");
+			}
+			else
+			{
+				header("Location: home/message=cannotunsubscribe");
+			}
 		}
 	}
